long -> long long enlargement of types to allow for left/rigth shift of 53 bits
git-svn-id: http://svn.osgeo.org/geos/trunk@186 5242fede-7e19-0410-aef8-94bd7d2200fb

+++ source/headers/indexQuadtree.h
@@ -50,7 +50,7 @@ public:
 	string toString();
 private:
 	double x;
-	long xBits;
+	long long xBits;
 };
 
 /**
